Add files via upload
The prevalence of liver diseases is rising due to factors such as excessive alcohol consumption, inhalation of harmful gases, consumption of contaminated food, pickles, and certain medications. Recognizing this alarming trend, particularly in India, where 18.3% of the global liver disease-related deaths occurred in 2015, there is an urgent need for early diagnosis to mitigate the socio-economic impact on patients and their families.

The aim of the project was to devise a prediction mechanism to determine the likelihood of an individual developing cirrhosis based on the presence of specific chemical compounds in the body, including bilirubin, albumin, proteins, and alkaline phosphatase. Early prediction is not only beneficial in disease management but also alleviates the load on healthcare professionals and diagnostic centers, while also potentially curtailing the socio-economic implications for affected families.

To achieve this, multiple predictive algorithms were evaluated on the dataset. The main objective was to optimize the Recall score, minimizing the risk of false negatives. After rigorous hyper-tuning, the Decision Tree model emerged as the most efficient, delivering the highest Recall score with rapid computation. Thus, this model offers promise in early detection, which could significantly impact disease management and socio-economic outcomes related to liver diseases.